gemini_assistant/requirements.txt
PIL
pillow
python-dotenv
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.138.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
google-pasta==0.2.0
googleapis-common-protos==1.63.2
pydantic==2.8.2
pydantic_core==2.20.1
pymongo==4.8.0
h11==0.14.0
uvicorn==0.30.6
anyio==4.4.0
fastapi==0.114.2

